Simplify the file even more
Although it was already pretty simple it could be made even smaller and
in my eyes easier to understand, ripped out the functionality that two
functions did and directly implemented them on their variables since we
were just returning them regardless.

Also renamed the function formatDiff into formatTime since that's all it
does now, added some comments to the top of the file and redid the
measures into an actual array instead of splitting them into an array.